services/user_ms/requirements.txt
Django==4.0
psycopg2-binary==2.9.2
djangorestframework==3.12.4
django-oauth-toolkit==1.5.0
pytz==2021.3